Move fragment logic inside activity
